Magento Open Source v2.0.2

+++ dev/tests/integration/testsuite/Magento/Catalog/_files/product_simple_multistore_rollback.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Copyright © 2016 Magento. All rights reserved.
+ * Copyright © 2015 Magento. All rights reserved.
  * See COPYING.txt for license details.
  */
 
@@ -10,14 +10,11 @@ $registry = \Magento\TestFramework\Helper\Bootstrap::getObjectManager()->get('Ma
 $registry->unregister('isSecureArea');
 $registry->register('isSecureArea', true);
 
-$repository = \Magento\TestFramework\Helper\Bootstrap::getObjectManager()->create(
-    'Magento\Catalog\Model\ProductRepository'
-);
-try {
-    $product = $repository->get('simple', false, null, true);
+/** @var $product \Magento\Catalog\Model\Product */
+$product = \Magento\TestFramework\Helper\Bootstrap::getObjectManager()->create('Magento\Catalog\Model\Product');
+$product->load(1);
+if ($product->getId()) {
     $product->delete();
-} catch (\Magento\Framework\Exception\NoSuchEntityException $e) {
-    //Entity already deleted
 }
 
 $registry->unregister('isSecureArea');
